+++ Client/PROJECTS/Jamtaba-common.pri
@@ -56,6 +56,7 @@ HEADERS += gui/PreferencesDialog.h
 HEADERS += gui/NinjamRoomWindow.h
 HEADERS += gui/BaseTrackView.h
 HEADERS += gui/NinjamTrackView.h
+HEADERS += gui/NinjamTrackGroupView.h
 HEADERS += gui/NinjamPanel.h
 HEADERS += gui/BusyDialog.h
 HEADERS += gui/ChatPanel.h
@@ -116,6 +117,7 @@ SOURCES += gui/NinjamRoomWindow.cpp
 SOURCES += gui/BaseTrackView.cpp
 SOURCES += audio/NinjamTrackNode.cpp
 SOURCES += gui/NinjamTrackView.cpp
+SOURCES += gui/NinjamTrackGroupView.cpp
 SOURCES += audio/MetronomeTrackNode.cpp
 SOURCES += gui/NinjamPanel.cpp
 SOURCES += ninjam/UserChannel.cpp
