+++ Client/src/Common/gui/NinjamTrackView.cpp
@@ -1,5 +1,4 @@
 #include "NinjamTrackView.h"
-#include "ui_TrackGroupView.h"
 #include "BaseTrackView.h"
 #include "ui_BaseTrackView.h"
 #include <QLineEdit>
@@ -27,10 +26,10 @@ NinjamTrackView::NinjamTrackView(Controller::MainController *mainController, lon
 
     setUnlightStatus(true);/** disabled/grayed until receive the first bytes. When the first bytes
     // are downloaded the 'on_channelXmitChanged' slot is executed and this track is enabled.*/
-
 }
 
-void NinjamTrackView::setInitialValues(const Persistence::CacheEntry &initialValues){
+void NinjamTrackView::setInitialValues(const Persistence::CacheEntry &initialValues)
+{
     cacheEntry = initialValues;
 
     // remember last track values
@@ -122,85 +121,4 @@ void NinjamTrackView::updateBoostValue()
     BaseTrackView::updateBoostValue();
     cacheEntry.setBoost(mainController->getTrackNode(getTrackID())->getBoost());
     mainController->getUsersDataCache()->updateUserCacheEntry(cacheEntry);
-}
-
-// ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-void NinjamTrackGroupView::updateGeoLocation()
-{
-    Geo::Location location = mainController->getGeoLocation(this->userIP);
-    countryLabel->setText(
-        "<img src=:/flags/flags/" + location.getCountryCode().toLower() +".png> <br>"
-        + location.getCountryName());
-}
-
-// ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-
-NinjamTrackGroupView::NinjamTrackGroupView(QWidget *parent,
-                                           Controller::MainController *mainController, long trackID,
-                                           QString channelName,
-                                           Persistence::CacheEntry initialValues) :
-    TrackGroupView(parent),
-    mainController(mainController),
-    userIP(initialValues.getUserIP())
-{
-    setSizePolicy(QSizePolicy(QSizePolicy::Fixed, QSizePolicy::Preferred));
-
-    // change the top panel layout to vertical (original is horizontal)
-    ui->topPanel->layout()->removeWidget(ui->groupNameField);
-    delete ui->topPanel->layout();
-    ui->topPanel->setLayout(new QVBoxLayout());
-    ui->topPanel->layout()->setContentsMargins(3, 6, 3, 3);
-
-    // replace the original QLineEdit with a MarqueeLabel
-
-    groupNameLabel = new MarqueeLabel();
-    delete ui->groupNameField;
-    groupNameLabel->setObjectName("groupNameField");
-    ui->topPanel->layout()->addWidget(groupNameLabel);
-
-    setGroupName(initialValues.getUserName());
-
-    // country flag label
-    countryLabel = new QLabel();
-    countryLabel->setObjectName("countryLabel");
-    countryLabel->setTextFormat(Qt::RichText);
-    updateGeoLocation();
-
-    ui->topPanel->layout()->addWidget(countryLabel);
-
-    // create the first subchannel by default
-    NinjamTrackView* newTrackView = addTrackView(trackID);
-    newTrackView->setChannelName(channelName);
-    newTrackView->setInitialValues(initialValues);
-}
-
-NinjamTrackView *NinjamTrackGroupView::createTrackView(long trackID)
-{
-    return new NinjamTrackView(mainController, trackID);
-}
-
-void NinjamTrackGroupView::setGroupName(const QString &groupName)
-{
-    groupNameLabel->setText(groupName);
-}
-
-void NinjamTrackGroupView::setNarrowStatus(bool narrow)
-{
-    foreach (BaseTrackView *trackView, trackViews) {
-        bool setToWide = !narrow && trackViews.size() <= 1;
-        if (setToWide)
-            trackView->setToWide();
-        else
-            trackView->setToNarrow();
-    }
-}
-
-void NinjamTrackGroupView::updateGuiElements()
-{
-    TrackGroupView::updateGuiElements();
-    groupNameLabel->updateMarquee();
-}
-
-NinjamTrackGroupView::~NinjamTrackGroupView()
-{
 }
